feature/ add exceptions

+++ apiWithDb/src/main/java/com/example/apiWithDb/service/impl/stuffmemberServiceimpl.java
@@ -1,5 +1,6 @@
 package com.example.apiWithDb.service.impl;
 
+import com.example.apiWithDb.exception.StuffMemberNotFoundException;
 import com.example.apiWithDb.model.StuffMember;
 import com.example.apiWithDb.repository.stuffMemberRepository;
 import com.example.apiWithDb.service.StuffMemberService;
@@ -30,12 +31,16 @@ public class stuffmemberServiceimpl implements StuffMemberService {
 
     @Override
     public String deleteStuffMember(String stuffmemberId) {
+        if(stuffmemberRepository.findById(stuffmemberId).isEmpty())
+            throw new StuffMemberNotFoundException("Запрошенный работник не существует!");
         stuffmemberRepository.deleteById(stuffmemberId);
         return "Success";
     }
 
     @Override
     public StuffMember getStuffMember(String stuffMemberId) {
+        if(stuffmemberRepository.findById(stuffMemberId).isEmpty())
+            throw new StuffMemberNotFoundException("Запрошенный работник не существует!");
         return stuffmemberRepository.findById(stuffMemberId).get();
     }
